Add Location / Address Book Flow
In the Address book view and the title of the location's annotation we
will need to optimize the data for the following:  1.) Brett's Home - If
a user's Address Book record has the following: a. First & Last Name b.
Home Address 2.) Brett's Work - If a user's Address Book record has the
following: a. First & Last Name b. Work Address 3.) UnitedWeego Inc. -
If a user's Address Book record has the following: a. Company Name b.
Work Address